add requirements

+++ run.py
@@ -89,8 +89,7 @@ def main(args):
     training_args = TrainingArguments(
         output_dir=args.output_dir,
         overwrite_output_dir=True,
-        fp16=True, 
-        # bf16=True, 
+        bf16=True, 
         # logging strategies
         logging_dir=f"{args.output_dir}/logs",
         logging_strategy="steps",
